requirements.txt
anyio==4.8.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
dotenv==0.9.9
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
python-dotenv==1.0.1
python-telegram-bot==22.0
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
